CHANGE:Added the previous value to features
Added the previous row's oil price as one of the features required to
predict the future price. Fixed InexactError where the number of samples
taken for both training and testing where not a whole number.

Started training the model based on 300 epochs, have also changed the
overall model to further expand upon the features provided.